feat: Update GKE Inference Gateway dashboard for older versions (#1197)
* feat: Update GKE Inference Gateway dashboard for older versions

The GKE Inference Gateway dashboard is updated to specify that it is for versions older than v1.1.0.

The documentation link in the prometheus metadata is also corrected.

* feat: Update GKE Inference Gateway dashboard for older versions

The GKE Inference Gateway dashboard is updated to specify that it is for versions older than v1.1.0.

The documentation link in the prometheus metadata is also corrected.